pyproject.toml
[tool.poetry]
name = "innocpc"
version = "0.1.0"
description = ""
authors = ["Mark Nicholson"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.2"
uvicorn = "^0.14.0"
python-dotenv = "^0.18.0"
pony = "^0.7.14"
psycopg2 = "^2.9.1"
black = "^21.6b0"
isort = "^5.9.2"
Jinja2 = "^3.0.1"
python-multipart = "^0.0.5"
PyJWT = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
